Reproduce namespace conflict
This commit shows that ETL job evaluation occurs in a way where the top level namespace is set to "Kiba" rather than nothing. This means that one can get conflicts if a non namespaced class of their own work has the same name as a class provided under the Kiba namespace.